update refs to stats redux submodule
