+++ src/redux/stats/thunks/index.ts
@@ -1,0 +1,1 @@
+export * from "./loadStatsData";
