+++ src/redux/types.ts
@@ -1,8 +1,8 @@
 import { AppState } from "@app/redux/app";
 import { NotificatorState } from "@app/redux/notificator";
+import { StatsState } from "@app/redux/stats";
 import { CommentsState } from "./comments/types";
 import { ModalState } from "./modal/types";
-import { StatsState } from "./stats/types";
 
 /** Redux action */
 export interface Action {
